requirements.txt
streamlit
pandas
requests
websocket-client
streamlit-autorefresh
python-dotenv